refactor: extract dom-result-reconcile.ts from dom-dispatch.ts (coherence 0.45)
Move result validation, lifecycle reconciliation, frame result picking, and
context enrichment into dom-result-reconcile.ts. dom-dispatch.ts retains
action execution and the main dispatcher.
