+++ src/background/dom-dispatch.ts
@@ -20,10 +20,18 @@ import { domPrimitiveWaitForStable, domPrimitiveActionDiff } from './dom-primiti
 import { domPrimitiveOverlay } from './dom-primitives-overlay.js'
 import { domPrimitiveIntent } from './dom-primitives-intent.js'
 import { isCDPEscalatable, tryCDPEscalation } from './cdp-dispatch.js'
-import { isReadOnlyAction, isMutatingAction } from './action-metadata.js'
+import { isReadOnlyAction } from './action-metadata.js'
 import { errorMessage } from '../lib/error-utils.js'
 import { delay } from '../lib/timeout-utils.js'
 import { normalizeFrameArg, resolveMatchedFrameIds } from './frame-targeting.js'
+import {
+  toDOMResult,
+  pickFrameResult,
+  mergeListInteractive,
+  deriveAsyncStatusFromDOMResult,
+  enrichWithEffectiveContext,
+  sendToastForResult
+} from './dom-result-reconcile.js'
 
 function parseDOMParams(query: PendingQuery): DOMActionParams | null {
   try {
@@ -31,19 +39,6 @@ function parseDOMParams(query: PendingQuery): DOMActionParams | null {
   } catch {
     return null
   }
-}
-
-function hasMatchedTargetEvidence(result: DOMResult): boolean {
-  const matched = result.matched
-  if (!matched || typeof matched !== 'object' || Array.isArray(matched)) return false
-  return (
-    typeof matched.selector === 'string' ||
-    typeof matched.tag === 'string' ||
-    typeof matched.element_id === 'string' ||
-    typeof matched.aria_label === 'string' ||
-    typeof matched.role === 'string' ||
-    typeof matched.text_preview === 'string'
-  )
 }
 
 type DOMExecutionTarget = { tabId: number; allFrames: true } | { tabId: number; frameIds: number[] }
@@ -59,80 +54,7 @@ async function resolveExecutionTarget(tabId: number, frame: unknown): Promise<DO
   return { tabId, frameIds }
 }
 
-/** Pick the best result from multi-frame executeScript. Prefers main frame, falls back to first success. */
-function pickFrameResult(results: chrome.scripting.InjectionResult[]): { result: unknown; frameId: number } | null {
-  const mainFrame = results.find((r) => r.frameId === 0)
-  if (mainFrame?.result && (mainFrame.result as DOMResult).success) {
-    return { result: mainFrame.result, frameId: 0 }
-  }
-  for (const r of results) {
-    if (r.result && (r.result as DOMResult).success) {
-      return { result: r.result, frameId: r.frameId }
-    }
-  }
-  if (mainFrame?.result) return { result: mainFrame.result, frameId: 0 }
-  return results[0] ? { result: results[0].result, frameId: results[0].frameId } : null
-}
-
-/** Merge list_interactive results from all frames (up to 100 elements). */
-function mergeListInteractive(results: chrome.scripting.InjectionResult[]): {
-  success: boolean
-  elements: unknown[]
-  candidate_count?: number
-  scope_rect_used?: unknown
-  error?: string
-  message?: string
-} {
-  const elements: unknown[] = []
-  let firstError: { error?: string; message?: string } | null = null
-  let firstScopeRectUsed: unknown
-  for (const r of results) {
-    const res = r.result as {
-      success?: boolean
-      elements?: unknown[]
-      scope_rect_used?: unknown
-      error?: string
-      message?: string
-    } | null
-    if (res?.success === false) {
-      if (!firstError) firstError = { error: res.error, message: res.message }
-      continue
-    }
-    if (firstScopeRectUsed === undefined && res?.scope_rect_used !== undefined) {
-      firstScopeRectUsed = res.scope_rect_used
-    }
-    if (res?.elements) elements.push(...res.elements)
-    if (elements.length >= 100) break
-  }
-  if (elements.length === 0 && firstError?.error) {
-    return { success: false, elements: [], error: firstError.error, message: firstError.message }
-  }
-  const cappedElements = elements.slice(0, 100)
-  const merged: {
-    success: boolean
-    elements: unknown[]
-    candidate_count?: number
-    scope_rect_used?: unknown
-  } = {
-    success: true,
-    elements: cappedElements,
-    candidate_count: cappedElements.length
-  }
-  if (firstScopeRectUsed !== undefined) {
-    merged.scope_rect_used = firstScopeRectUsed
-  }
-  return merged
-}
-
 const WAIT_FOR_POLL_INTERVAL_MS = 80
-
-function toDOMResult(value: unknown): DOMResult | null {
-  if (!value || typeof value !== 'object') return null
-  const candidate = value as DOMResult
-  if (typeof candidate.success !== 'boolean') return null
-  if (typeof candidate.action !== 'string' || typeof candidate.selector !== 'string') return null
-  return candidate
-}
 
 /** Resolve which DOM action name to dispatch for wait_for based on params.
  *  Callers must validate mutual exclusivity before calling this. */
@@ -360,106 +282,6 @@ async function executeIntentAction(
 const STABILITY_ACTIONS = new Set(['wait_for_stable', 'action_diff'])
 const OVERLAY_ACTIONS = new Set(['dismiss_top_overlay', 'auto_dismiss_overlays'])
 const INTENT_ACTIONS = new Set(['open_composer', 'submit_active_composer', 'confirm_top_dialog'])
-
-function sendToastForResult(
-  tabId: number,
-  readOnly: boolean,
-  result: { success?: boolean; error?: string },
-  actionToast: ActionToastFn,
-  toastLabel: string,
-  toastDetail: string | undefined
-): void {
-  if (readOnly) return
-  if (result.success) {
-    actionToast(tabId, toastLabel, toastDetail, 'success')
-  } else {
-    actionToast(tabId, toastLabel, result.error || 'failed', 'error')
-  }
-}
-
-function reconcileDOMLifecycle(
-  action: string,
-  selector: string,
-  result: unknown
-): { result: unknown; status: 'complete' | 'error'; error?: string } {
-  const domResult = toDOMResult(result)
-  if (!domResult) {
-    if (!isMutatingAction(action)) return { result, status: 'complete' }
-    const coerced: DOMResult = {
-      success: false,
-      action,
-      selector,
-      error: 'status_mismatch',
-      message: `Mutating action returned non-DOM payload: ${action}`
-    }
-    return { result: coerced, status: 'error', error: 'status_mismatch' }
-  }
-
-  if (!domResult.success) {
-    return {
-      result: domResult,
-      status: 'error',
-      error: domResult.error || domResult.message || 'dom_action_failed'
-    }
-  }
-
-  if (domResult.error) {
-    const coerced: DOMResult = {
-      ...domResult,
-      success: false,
-      error: 'status_mismatch',
-      message: `Payload marked success but includes error: ${domResult.error}`
-    }
-    return { result: coerced, status: 'error', error: 'status_mismatch' }
-  }
-
-  if (isMutatingAction(action) && !hasMatchedTargetEvidence(domResult)) {
-    const coerced: DOMResult = {
-      ...domResult,
-      success: false,
-      error: 'missing_match_evidence',
-      message: `Mutating action completed without matched target evidence: ${action}`
-    }
-    return { result: coerced, status: 'error', error: 'missing_match_evidence' }
-  }
-
-  return { result: domResult, status: 'complete' }
-}
-
-function deriveAsyncStatusFromDOMResult(
-  action: string,
-  selector: string,
-  result: unknown
-): { result: unknown; status: 'complete' | 'error'; error?: string } {
-  const reconciled = reconcileDOMLifecycle(action, selector, result)
-  if (reconciled.status === 'complete') {
-    return reconciled
-  }
-  return {
-    status: 'error',
-    error: reconciled.error || 'dom_action_failed',
-    result: reconciled.result
-  }
-}
-
-// Enrich results with effective tab context (post-execution URL).
-// Agents compare resolved_url (dispatch time) vs effective_url (execution time) to detect drift.
-async function enrichWithEffectiveContext(tabId: number, result: unknown): Promise<unknown> {
-  try {
-    const tab = await chrome.tabs.get(tabId)
-    if (result && typeof result === 'object' && !Array.isArray(result)) {
-      return {
-        ...(result as Record<string, unknown>),
-        effective_tab_id: tabId,
-        effective_url: tab.url,
-        effective_title: tab.title
-      }
-    }
-    return result
-  } catch {
-    return result
-  }
-}
 
 // #lizard forgives
 export async function executeDOMAction(
